Move FakeCreditCardService out of todo/
It was misplaced there since nothing in todo/ used it
