+++ tests/Integration/UseCases/CreditCardPaymentNotification/CreditCardNotificationUseCaseTest.php
@@ -20,7 +20,7 @@ use WMDE\Fundraising\Frontend\DonationContext\Tests\Fixtures\SucceedingDonationA
 use WMDE\Fundraising\Frontend\DonationContext\Tests\Integration\DonationEventLoggerAsserter;
 use WMDE\Fundraising\Frontend\DonationContext\UseCases\CreditCardPaymentNotification\CreditCardNotificationUseCase;
 use WMDE\Fundraising\Frontend\DonationContext\UseCases\CreditCardPaymentNotification\CreditCardPaymentHandlerException;
-use WMDE\Fundraising\Frontend\Tests\Fixtures\FakeCreditCardService;
+use WMDE\Fundraising\Frontend\PaymentContext\Infrastructure\FakeCreditCardService;
 use WMDE\Fundraising\Frontend\Tests\Fixtures\ThrowingEntityManager;
 
 /**
